Fixed comilation on x86_64
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@311379 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/math.c
@@ -608,7 +608,7 @@ PHP_FUNCTION(pow)
 
 		/* calculate pow(long,long) in O(log exp) operations, bail if overflow */
 		while (i >= 1) {
-			int overflow;
+			long overflow;
 			double dval = 0.0;
 
 			if (i % 2) {
